+++ test/integration/standalone/multi_segments/multi_segments_suite_test.go
@@ -1,0 +1,109 @@
+// Licensed to Apache Software Foundation (ASF) under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Apache Software Foundation (ASF) licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package integration_query_test
+
+import (
+	"testing"
+	"time"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/gleak"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
+
+	"github.com/apache/skywalking-banyandb/pkg/grpchelper"
+	"github.com/apache/skywalking-banyandb/pkg/logger"
+	"github.com/apache/skywalking-banyandb/pkg/pool"
+	"github.com/apache/skywalking-banyandb/pkg/test/flags"
+	"github.com/apache/skywalking-banyandb/pkg/test/gmatcher"
+	"github.com/apache/skywalking-banyandb/pkg/test/helpers"
+	"github.com/apache/skywalking-banyandb/pkg/test/setup"
+	"github.com/apache/skywalking-banyandb/pkg/timestamp"
+	test_cases "github.com/apache/skywalking-banyandb/test/cases"
+	casesmeasure "github.com/apache/skywalking-banyandb/test/cases/measure"
+	casesstream "github.com/apache/skywalking-banyandb/test/cases/stream"
+	casestopn "github.com/apache/skywalking-banyandb/test/cases/topn"
+	casestrace "github.com/apache/skywalking-banyandb/test/cases/trace"
+	integration_standalone "github.com/apache/skywalking-banyandb/test/integration/standalone"
+)
+
+func TestIntegrationMultiSegments(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Integration Multi Segments Suite", Label(integration_standalone.Labels...))
+}
+
+var (
+	connection *grpc.ClientConn
+	now        time.Time
+	baseTime   time.Time
+	deferFunc  func()
+	goods      []gleak.Goroutine
+)
+
+var _ = SynchronizedBeforeSuite(func() []byte {
+	goods = gleak.Goroutines()
+	Expect(logger.Init(logger.Logging{
+		Env:   "dev",
+		Level: flags.LogLevel,
+	})).To(Succeed())
+	var addr string
+	addr, _, deferFunc = setup.Standalone()
+	ns := timestamp.NowMilli().UnixNano()
+	now = time.Unix(0, ns-ns%int64(time.Minute))
+	baseTime = time.Date(now.Year(), now.Month(), now.Day(), 0o0, 0o2, 0, 0, now.Location())
+	test_cases.Initialize(addr, baseTime)
+	return []byte(addr)
+}, func(address []byte) {
+	var err error
+	connection, err = grpchelper.Conn(string(address), 10*time.Second,
+		grpc.WithTransportCredentials(insecure.NewCredentials()))
+	casesstream.SharedContext = helpers.SharedContext{
+		Connection: connection,
+		BaseTime:   baseTime,
+	}
+	casesmeasure.SharedContext = helpers.SharedContext{
+		Connection: connection,
+		BaseTime:   baseTime,
+	}
+	casestopn.SharedContext = helpers.SharedContext{
+		Connection: connection,
+		BaseTime:   baseTime,
+	}
+	casestrace.SharedContext = helpers.SharedContext{
+		Connection: connection,
+		BaseTime:   baseTime,
+	}
+	Expect(err).NotTo(HaveOccurred())
+})
+
+var _ = SynchronizedAfterSuite(func() {
+	if connection != nil {
+		Expect(connection.Close()).To(Succeed())
+	}
+}, func() {})
+
+var _ = ReportAfterSuite("Integration Query Suite", func(report Report) {
+	if report.SuiteSucceeded {
+		if deferFunc != nil {
+			deferFunc()
+		}
+		Eventually(gleak.Goroutines, flags.EventuallyTimeout).ShouldNot(gleak.HaveLeaked(goods))
+		Eventually(pool.AllRefsCount, flags.EventuallyTimeout).Should(gmatcher.HaveZeroRef())
+	}
+})