+++ CHANGES.md
@@ -59,6 +59,7 @@ Release Notes.
 - Fix returning empty result when using IN operatior on the array type tags.
 - Fix memory leaks and OOM issues in streaming processing by implementing deduplication logic in priority queues and improving sliding window memory management.
 - Fix etcd prefix matching any key that starts with this prefix.
+- Fix the sorting timestamps issue of the measure model when there are more than one segment.
 
 ### Document